Checkin of fix for assertion failure exposed by Mark Miller's testhdf5 benchmark.
In essence, the problem was failure to set appropriate tags when protecting
entries in the fractal heap sanity check code. This resulted in H5AC protect
failing when H5C_DO_TAGGING_SANITY_CHECKS is set.

This bug was only detected in Mark Miller's bench mark as the existing test 
code did not generate the cache contents necessary to execute the pathway in 
which the failure occurred.

Cleaned up two other possible occurrences of the tagging issue in passing.

Also cleaned up several over sensitive asserts exposed by the fix.

Note that this checkin does NOT include code to verify the fix.  

Verified the fix manually with Mark Miller's benchmark.

Tested: Serial/debug on Jam
        Serail/debug and Parallel/debug on Mercury